Update app.js
These changes fix the bug where the UI was not updating on the first search, returning undefined, then subsequently displaying results previously searched for and not the currently searched for query. 
It also brings about an alert for the user to enter a valid city in-case the API returns a 404 city not found.